Datatype of minimumSamplingInterval should be "Duration" (i.e Double)

+++ test/server/test_subscription.js
@@ -1317,7 +1317,7 @@ describe("Subscription#adjustSamplingInterval", function () {
         readAttribute: function(attributeId) {
 
             attributeId.should.eql(AttributeIds.MinimumSamplingInterval);
-            return  new DataValue({value: {dataType: DataType.UInt32, value: 0 }});
+            return  new DataValue({value: {dataType: DataType.Double, value: 0.0 }});
         }
     };
